pyproject.toml
[project]
name = "multi-agent-networking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "langgraph-supervisor>=0.0.21",
    "python-dotenv>=1.1.0",
]
